# swagger/swagger.yaml
basePath: /api/
consumes:
- application/json
definitions:
...
  AuthorAPIModel:
    properties:
      count:
        description: The popularity index of the author
        example: 1111
        format: int64
        type: integer
        x-go-name: Count
      hasIcelandicQuotes:
        description: Whether or not this author has some icelandic quotes
        example: true
        type: boolean
        x-go-name: HasIcelandicQuotes
      id:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: Id
      name:
        description: Name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      nrOfEnglishQuotes:
        description: How many quotes in English this author has
        example: 78
        format: int64
        type: integer
        x-go-name: NrOfEnglishQuotes
      nrOfIcelandicQuotes:
        description: How many quotes in Icelandic this author has
        example: 6
        format: int64
        type: integer
        x-go-name: NrOfIcelandicQuotes
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
...
  TopicViewAPIModel:
    properties:
      authorId:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: AuthorId
      isIcelandic:
        description: Whether or not this quote is in Icelandic or not
        example: false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote
        example: Float like a butterfly, sting like a bee.
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: QuoteId
      topicId:
        description: The topic's id (if topic id / name not supplied this will return
          a zero id)
        example: 10
        format: int64
        type: integer
        x-go-name: TopicId
      topicName:
        description: The topic's name (if topic id / name not supplied this will return
          empty string "")
        example: inspirational
        type: string
        x-go-name: TopicName
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
  UserResponse:
    properties:
      apiKey:
        description: The api-key that the user should send to get access to the api
        example: 1d8db1d2-6f5b-4254-8b74-44f5e5229add
        type: string
        x-go-name: ApiKey
      id:
        description: The user id
        example: 1
        format: int64
        type: integer
        x-go-name: Id
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
...
host: quotel-api.com
info:
  contact:
    email: skjaldbaka17@gmail.com
    name: Þórður Ágústsson
  description: |-
    Quotel API offers a complete feature rich REST API access to all your quotes needs. This is the documentation of our quotes API.
    To use it you first need to sign up here https://somethingsomethingdarkside.com to get an authentication api key.
  title: quotes-api.
  version: 1.0.0
paths:
...
  /search:
    post:
      description: Search for quotes / authors by a general string-search that searches
        both in the names of the authors and the quotes themselves
      operationId: SearchByString
      parameters:
      - description: The structure of the request for searching quotes/authors
        in: body
        name: Body
        required: true
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              description: The particular language that the quote should be in
              example: English
              type: string
              x-go-name: Language
            page:
              description: The page you are asking for, starts with 0.
              example: 0
              format: int64
              minimum: 0
              type: integer
              x-go-name: Page
            pageSize:
              default: 25
              description: The number of quotes to be returned on each "page"
              example: 30
              format: int64
              maximum: 200
              minimum: 1
              type: integer
              x-go-name: PageSize
            searchString:
              description: The string to be used in the search
              example: sting like butterfly
              type: string
              x-go-name: SearchString
          required:
          - apiKey
          - searchString
          type: object
      responses:
        "200":
          $ref: '#/responses/topicViewsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - SEARCH
  /search/authors:
    post:
      description: Authors search. Searching authors by a given search string
      operationId: SearchAuthorsByString
      parameters:
      - description: The structure of the request for searching quotes/authors
        in: body
        name: Body
        required: true
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              description: The particular language that the quote should be in
              example: English
              type: string
              x-go-name: Language
            page:
              description: The page you are asking for, starts with 0.
              example: 0
              format: int64
              minimum: 0
              type: integer
              x-go-name: Page
            pageSize:
              default: 25
              description: The number of quotes to be returned on each "page"
              example: 30
              format: int64
              maximum: 200
              minimum: 1
              type: integer
              x-go-name: PageSize
            searchString:
              description: The string to be used in the search
              example: Ali Muhammad
              type: string
              x-go-name: SearchString
          required:
          - apiKey
          - searchString
          type: object
      responses:
        "200":
          $ref: '#/responses/authorsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - SEARCH
  /search/quotes:
    post:
      description: Quotes search. Searching quotes by a given search string
      operationId: SearchQuotesByString
      parameters:
      - description: The structure of the request for searching quotes/authors
        in: body
        name: Body
        required: true
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              description: The particular language that the quote should be in
              example: English
              type: string
              x-go-name: Language
            page:
              description: The page you are asking for, starts with 0.
              example: 0
              format: int64
              minimum: 0
              type: integer
              x-go-name: Page
            pageSize:
              default: 25
              description: The number of quotes to be returned on each "page"
              example: 30
              format: int64
              maximum: 200
              minimum: 1
              type: integer
              x-go-name: PageSize
            searchString:
              description: The string to be used in the search
              example: sting like butterfly
              type: string
              x-go-name: SearchString
          required:
          - apiKey
          - searchString
          type: object
      responses:
        "200":
          $ref: '#/responses/topicViewsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - SEARCH
...
  /users/login:
    post:
      description: Login to get the apiKey for the user
      operationId: Login
      parameters:
      - description: The structure of the Login request
        in: body
        name: Body
        schema:
          properties:
            email:
              description: The email for the user you want to login
              example: example@gmail.com
              type: string
              x-go-name: Email
            password:
              description: The password for the user
              example: "1234567890"
              type: string
              x-go-name: Password
          type: object
      responses:
        "200":
          $ref: '#/responses/userResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "401":
          $ref: '#/responses/incorrectCredentialsResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - USERS
  /users/signup:
    post:
      description: Create A user to get a free ApiKey
      operationId: SignUp
      parameters:
      - description: The structure of the sign up request
        in: body
        name: Body
        schema:
          properties:
            email:
              description: The email for the created user (used for validation and
                password retrieval)
              example: example@gmail.com
              type: string
              x-go-name: Email
            name:
              description: The name of the creator
              example: Robert Huldars
              type: string
              x-go-name: Name
            password:
              description: The password for the user, used to get the ApiKey and look
                at stats (coming soon). Must be a relatively strong pass
              example: "1234567890"
              type: string
              x-go-name: Password
            passwordConfirmation:
              description: The confirmation for the password. Must be same as "password"
              example: "1234567890"
              type: string
              x-go-name: PasswordConfirmation
          type: object
      responses:
        "200":
          $ref: '#/responses/userResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - USERS
produces:
- application/json
responses:
...
  authorsResponse:
    description: Data structure representing the response for authors
    schema:
      items:
        $ref: '#/definitions/AuthorAPIModel'
      type: array
  incorrectBodyStructureResponse:
    description: Data structure representing the error response to a wrongly structured
      request body
    schema:
      properties:
        message:
          description: The error message
          example: request body is not structured correctly.
          type: string
          x-go-name: Message
      type: object
  incorrectCredentialsResponse:
    description: Data structure representing the error response to an incorrect Credentials
      error
    schema:
      properties:
        message:
          description: The error message
          example: Valar Dohaeris
          type: string
          x-go-name: Message
      type: object
  internalServerErrorResponse:
    description: Data structure representing the error response to an internal server
      error
    schema:
      properties:
        message:
          description: The error message
          example: Please try again later.
          type: string
          x-go-name: Message
      type: object
...
  topicViewsResponse:
    description: Data structure representing the response for a quote based on a particular
      topic
    schema:
      items:
        $ref: '#/definitions/TopicViewAPIModel'
      type: array
...
  userResponse:
    description: Data structure representing a user response
    schema:
      $ref: '#/definitions/UserResponse'
schemes:
- http
securityDefinitions:
  apikey:
    description: To use the api you need an API key. The basic API key lets you make
      up to 100 requests an hour.
    in: body
    name: apiKey
    type: apiKey
swagger: "2.0"
...